refactor: add timesheets repository layer and move code from service there
