+++ plugins/services.ts
@@ -9,8 +9,11 @@ import WorkSchemeService from '~/services/work-scheme-service';
 import TimesheetsService from '~/services/timesheets-service';
 import ContractsService from '~/services/contracts-service';
 import BridgeService from '~/services/bridge-service';
+import TimesheetsRepository from '~/api/timesheets/timesheets-repository';
 
 export default defineNuxtPlugin(({$fire, $fireModule, $axios}, inject) => {
+  const timesheetsRepository = new TimesheetsRepository($fire);
+
   inject('contractsService', new ContractsService($axios));
   inject('customersService', new CustomersService($fire, $fireModule));
   inject('timeRecordsService', new TimeRecordsService($fire, $axios));
@@ -18,6 +21,6 @@ export default defineNuxtPlugin(({$fire, $fireModule, $axios}, inject) => {
   inject('employeesService', new EmployeesService($fire, $fireModule));
   inject('teamsService', new TeamsService($fire, $fireModule));
   inject('workSchemeService', new WorkSchemeService($axios));
-  inject('timesheetsService', new TimesheetsService($fire));
+  inject('timesheetsService', new TimesheetsService(timesheetsRepository));
   inject('bridgeService', new BridgeService($axios));
 });
